Updates calculation logic
Allows calculations to be chained and calculated. Follows the correct
order of operations. Also allows for previous result to be used in new
calculation. Added in additional logic to handle various input
combinations and corrected any related issues/bugs.

In process of adding in decimal point button and delete/backspace
functionality.